Handle lightning balance and error message

+++ Plugins/BTCPayServer.Plugins.NairaCheckout/ViewModels/MavapayVM/CreateQuoteResponseVm.cs
@@ -11,8 +11,8 @@ public class CreateQuoteResponseVm
     public string targetCurrency { get; set; }
     public int transactionFeesInSourceCurrency { get; set; }
     public int transactionFeesInTargetCurrency { get; set; }
-    public int amountInSourceCurrency { get; set; }
-    public int amountInTargetCurrency { get; set; }
+    public long amountInSourceCurrency { get; set; }
+    public long amountInTargetCurrency { get; set; }
     public string paymentMethod { get; set; }
     public DateTime expiry { get; set; }
     public bool isValid { get; set; }
@@ -47,14 +47,14 @@ public class CreatePayoutResponseModel
     public string targetCurrency { get; set; }
     public int transactionFeesInSourceCurrency { get; set; }
     public int transactionFeesInTargetCurrency { get; set; }
-    public int amountInSourceCurrency { get; set; }
-    public int amountInTargetCurrency { get; set; }
+    public long amountInSourceCurrency { get; set; }
+    public long amountInTargetCurrency { get; set; }
     public string paymentMethod { get; set; }
     public DateTime expiry { get; set; }
     public bool isValid { get; set; }
     public string invoice { get; set; }
     public string hash { get; set; }
-    public int totalAmountInSourceCurrency { get; set; }
+    public long totalAmountInSourceCurrency { get; set; }
     public int customerInternalFee { get; set; }
     public string orderId { get; set; }
     public string ErrorMessage { get; set; }
